Selectors used. StoreDevtoolsModule added

+++ src/app/components/task/task-clear/task-clear.component.ts
@@ -3,9 +3,8 @@ import { Store } from '@ngrx/store';
 import { map, catchError, of, finalize } from 'rxjs';
 import { TaskHttpService } from 'src/app/http/task.http.service';
 import { Task } from 'src/app/models/task';
-import { SpinnerService } from 'src/app/services/spinner.service';
-import { TaskState } from 'src/app/store';
-import { removeAllTask } from 'src/app/store/actions';
+import { TaskStoreState, getTasks } from 'src/app/store';
+import { removeAllTask, setLoading } from 'src/app/store/actions';
 
 @Component({
   selector: 'app-task-clear',
@@ -14,22 +13,22 @@ import { removeAllTask } from 'src/app/store/actions';
 export class TaskClearComponent {
   tasks: Task[] = [];
 
-  constructor(
-    private store: Store<{ taskReducer: TaskState }>,
-    private taskHttpService: TaskHttpService,
-    private spinnerService: SpinnerService
-  ) {
-    store.select('taskReducer').subscribe(state => (this.tasks = state.tasks));
+  constructor(private store: Store<TaskStoreState>, private taskHttpService: TaskHttpService) {
+    store.select(getTasks).subscribe(tasks => (this.tasks = tasks));
   }
 
   handleRemoveTask() {
-    this.spinnerService.setLoading(true);
+    if (!confirm('Are you sure you want to clear all tasks?')) {
+      return;
+    }
+
+    this.store.dispatch(setLoading({ loading: true }));
     this.taskHttpService
       .removeAll(this.tasks.map(i => i._id))
       .pipe(
         map(() => this.store.dispatch(removeAllTask())),
         catchError(err => of(alert(err.message))),
-        finalize(() => this.spinnerService.setLoading(false))
+        finalize(() => this.store.dispatch(setLoading({ loading: false })))
       )
       .subscribe();
   }
